* rc_io.c (krb5_rc_io_creat): Cast getpid() to int.         (Use of a long might overflow buffer if pid_t is 64 bits */
git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@12480 dc483132-0cff-0310-8789-dd5450dbe970

+++ src/lib/krb5/rcache/rc_io.c
@@ -101,7 +101,7 @@ krb5_error_code krb5_rc_io_creat (context, d, fn)
    if (fn)
      if (!(*fn = malloc(35)))
       { FREE(d->fn); return KRB5_RC_IO_MALLOC; }
-   (void) sprintf(d->fn,"%s%skrb5_RC%d",dir,PATH_SEPARATOR,UNIQUE);
+   (void) sprintf(d->fn,"%s%skrb5_RC%d",dir,PATH_SEPARATOR, (int) UNIQUE);
    c = d->fn + strlen(d->fn);
    (void) strcpy(c,"aaa");
    while ((d->fd = THREEPARAMOPEN(d->fn,O_WRONLY|O_CREAT|O_TRUNC|O_EXCL|O_BINARY,0600)) == -1)
